Update theming file structure and color naming
