+++ tv-shows/src/styles/theme/foundations/colors.ts
@@ -1,0 +1,11 @@
+const colors = {
+  white: "#FFFFFF",
+  purple: {
+    100: "#8D5CE5",
+    300: "#371687",
+    500: "#1B004C",
+  },
+  error: "#FF2498",
+};
+
+export default colors;
