+++ tv-shows/src/styles/theme/foundations/fontSizes.ts
@@ -1,0 +1,14 @@
+const fontSizes = {
+  xs: "12px",
+  sm: "14px",
+  md: "16px",
+  lg: "18px",
+  xl: "20px",
+  "2xl": "24px",
+  "3xl": "36px",
+  "4xl": "40px",
+  "5xl": "48px",
+  "6xl": "64px",
+};
+
+export default fontSizes;
