+++ tv-shows/src/styles/theme/foundations/fontWeights.ts
@@ -1,0 +1,7 @@
+const fontWeights = {
+  normal: 400,
+  medium: 500,
+  bold: 700,
+};
+
+export default fontWeights;
