+++ tv-shows/src/styles/theme/foundations/fonts.ts
@@ -1,0 +1,6 @@
+const fonts = {
+  body: "Roboto, sans-serif",
+  heading: "Roboto, sans-serif",
+};
+
+export default fonts;
